+++ CHANGELOG.md
@@ -1,5 +1,13 @@
 # ChangeLog
 
+## [1.0.2] 2020-04-09
+
+### Added
+
+- ParamKeyData/ParamKeyDisplay/ParamKeyDevice/ParamKeyProtocol/ParamKeyInterface/ParamKeyAction/ParamKeyTemplate
+- Query/Add/Modify/Remove for Cell Storage
+- Query/Get/Create/Modify/Delete for System Template
+
 ## [1.0.1] 2020-01-02
 ### Added